+++ system/src/Cms/Page.php
@@ -4,16 +4,17 @@ declare(strict_types=1);
 
 namespace Zolinga\System\Cms;
 
-use Zolinga\System\Events\{ListenerInterface,ContentEvent};
+use Zolinga\System\Events\Content\HtmlContentEvent;
+use Zolinga\System\Events\ListenerInterface;
 
 class Page implements ListenerInterface
 {
-    public function onContent(ContentEvent $event): void
+    public function onContent(HtmlContentEvent $event): void
     {
         global $api;
 
         // Page was already served by other CMS
-        if ($event->status != ContentEvent::STATUS_UNDETERMINED) {
+        if ($event->status != HtmlContentEvent::STATUS_UNDETERMINED) {
             return;
         }
         $html = file_get_contents('module://system/data/welcome.html') ?: '<h1>Welcome to Zolinga CMS</h1>';
@@ -28,6 +29,6 @@ class Page implements ListenerInterface
         $html = str_replace(array_map(fn ($k) => '{{'.$k.'}}', array_keys($vars)), array_values($vars), $html);
 
         $event->setContentHTML($html);
-        $event->setStatus(ContentEvent::STATUS_OK, 'Page served by Zolinga CMS');
+        $event->setStatus(HtmlContentEvent::STATUS_OK, 'Page served by Zolinga CMS');
     }
 }